# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Xendit Coding Test",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:8010",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "health"
        ],
        "description": "To get current service health status",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/rides": {
      "post": {
        "tags": [
          "ride"
        ],
        "description": "To update information of current rides",
        "parameters": [
          {
            "name": "createRide",
            "in": "body",
            "description": "Create new ride payload",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateRide"
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/CreateRideResponse"
            },
            "description": "Success create rider response."
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "get": {
        "tags": [
          "ride"
        ],
        "description": "To retrieve information about multiple rides",
        "parameters": [
          {
            "name": "perPage",
            "in": "query",
            "type": "string"
          },
          {
            "name": "pageNo",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success get list of rides.",
            "schema": {
              "$ref": "#/definitions/RidesResponse"
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/rides/{id}": {
      "get": {
        "tags": [
          "ride"
        ],
        "description": "To retrieve information about specific rides",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success get list of rides.",
            "schema": {
              "$ref": "#/definitions/RideResponse"
            }
          },
          "404": {
            "description": "Ride not found.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/Error"
            },
            "description": "Internal server error."
          }
        }
      }
    }
  },
  "definitions": {
    "Error": {
      "type": "object",
      "properties": {
        "error_code": {
          "type": "string",
          "example": "SERVER_ERROR"
        },
        "message": {
          "type": "string",
          "example": "Unknown error"
        }
      }
    },
    "CreateRide": {
      "type": "object",
      "properties": {
        "start_lat": {
          "type": "number",
          "example": 89.191
        },
        "start_long": {
          "type": "number",
          "example": 170.7893
        },
        "end_lat": {
          "type": "number",
          "example": 89.191
        },
        "end_long": {
          "type": "number",
          "example": 170.7893
        },
        "rider_name": {
          "type": "string",
          "example": "Jhon"
        },
        "driver_name": {
          "type": "string",
          "example": "Doe"
        },
        "driver_vehicle": {
          "type": "string",
          "example": "Yamaha"
        }
      }
    },
    "RideResponse": {
      "type": "object",
      "properties": {
        "startLat": {
          "type": "number",
          "example": 89.191
        },
        "startLong": {
          "type": "number",
          "example": 170.7893
        },
        "endLat": {
          "type": "number",
          "example": 89.191
        },
        "endLong": {
          "type": "number",
          "example": 170.7893
        },
        "rider_name": {
          "type": "string",
          "example": "Jhon"
        },
        "driver_name": {
          "type": "string",
          "example": "Doe"
        },
        "driver_vehicle": {
          "type": "string",
          "example": "Yamaha"
        }
      }
    },
    "RidesResponse": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "startLat": {
            "type": "number",
            "example": 89.191
          },
          "startLong": {
            "type": "number",
            "example": 170.7893
          },
          "endLat": {
            "type": "number",
            "example": 89.191
          },
          "endLong": {
            "type": "number",
            "example": 170.7893
          },
          "rider_name": {
            "type": "string",
            "example": "Jhon"
          },
          "driver_name": {
            "type": "string",
            "example": "Doe"
          },
          "driver_vehicle": {
            "type": "string",
            "example": "Yamaha"
          }
        }
      }
    },
    "CreateRideResponse": {
      "type": "object",
      "properties": {
        "rideID": {
          "type": "number",
          "example": 23
        },
        "startLat": {
          "type": "number",
          "example": 0.7893
        },
        "startLong": {
          "type": "number",
          "example": 113.9213
        },
        "endLat": {
          "type": "number",
          "example": 103.89
        },
        "endLong": {
          "type": "number",
          "example": 103.89
        },
        "riderName": {
          "type": "string",
          "example": "Jhon"
        },
        "driverName": {
          "type": "string",
          "example": "Doe"
        },
        "driverVehicle": {
          "type": "string",
          "example": "Suzuki"
        },
        "created": {
          "type": "string",
          "example": "2022-09-05 14:42:17"
        }
      }
    }
  }
}